cancel listing navigation

+++ app/src/main/java/com/gruppe16/birdwatcher/fragments/CreateItemFragment.kt
@@ -52,7 +52,10 @@ class CreateItemFragment : Fragment() {
         }
 
         binding.cancelListing.setOnClickListener {
-            db.deletePhotoFromFirebase(pictureId)
+            if (db.pictureUrl == null) {
+                db.deletePhotoFromFirebase(pictureId)
+            }
+            findNavController().navigate(R.id.action_createItem_to_homeFragment)
         }
 
         binding.saveListing.setOnClickListener {
